+++ lightwood/mixer/qclassic.py
@@ -9,7 +9,7 @@ from lightwood.api.types import TimeseriesSettings
 class QClassic(Neural):
     # wrapper class to be combined with Neural class when performance stabilizes
     def __init__(
-            self, stop_after: int, target: str, dtype_dict: Dict[str, str],
+            self, stop_after: float, target: str, dtype_dict: Dict[str, str],
             input_cols: List[str],
             timeseries_settings: TimeseriesSettings, target_encoder: BaseEncoder, net: str, fit_on_dev: bool,
             search_hyperparameters: bool):
